changed endpoint path according to task convention
