+++ src/routes/routes.ts
@@ -4,13 +4,11 @@ import * as productController from '../controllers/productsController.js';
 import * as phonesController from '../controllers/phonesController.js';
 import * as tabletsController from '../controllers/tabletsController.js';
 import * as accessoriesController from '../controllers/accessoriesController.js';
-import * as homeController from '../controllers/homeController.js';
 
 const router = express.Router();
 
-router.get('/home/hotprice', homeController.getHotPrice);
-router.get('/home/newproducts', homeController.getNewestProducts);
-
+router.get('/products/discount', productController.getHotPrice);
+router.get('/products/new', productController.getNewestProducts);
 router.get('/products/:category', productController.getAllByCategory);
 router.get('/products/:id/recommended', productController.getRecomended);
 router.get('/products', productController.getAll);
